requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
matplotlib==3.8.0
tqdm==4.66.1
denseweight==0.1.2
pillow==10.0.0
scikit-learn==1.3.0
scipy==1.9.3               # Updated to be compatible with Python 3.11
h5py==3.9.0
pandas==2.1.0
fsspec==2022.5.0
blosc2==2.0.0
cython>=0.29.21
FuzzyTM>=0.4.0
google-auth-oauthlib==0.5.0
tensorboard-data-server==0.7.0
protobuf==3.20.3
urllib3==1.26.15

